requirements.txt
matplotlib
seaborn
tensorflow
tqdm
keras
pillow
werkzeug
scikit-learn